病史添加用药史

+++ core/src/main/java/com/wxsoft/fcare/core/data/entity/previoushistory/History2.kt
@@ -1,3 +1,3 @@
 package com.wxsoft.fcare.core.data.entity.previoushistory
 
-data class History2(val drugId:String, val drugName:String,val dose:Float,val medicalHistoryId:String)
+data class History2(val drugId:String, val drugName:String,val dose:Int,val medicalHistoryId:String)
